Add polling example, document the others
