+++ examples/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod _1_simple_side_effects;
-pub mod _2_simple_order_of_effect;
-pub mod _3_complex_order_of_effects;
-pub mod _4_sequence_match;
-pub mod _5_setting_up_mock_values;
+pub mod _2_setting_up_mock_values;
+pub mod _3_simple_order_of_effect;
+pub mod _4_complex_order_of_effects;
+pub mod _5_sequence_match;
+pub mod _6_mocking_value_read_sequences_using_read_callback;
+pub mod _7_handling_polling_in_mocks;
 pub mod common;
